ao_log: Change flight type to uint16_t, widen APIs dealing with flight numbers to int32_t
Real Flight numbers are 16-bit unsigned values, but APIs using them
pass values that areoverloaded to also represent broken log slots
using negative numbers. This means that these APIs need to be able to
represent all 65535 possible flight numbers *and* all possible log
slots as negative values. Switch the variables holding only flight
numbers to uint16_t and those holding either a flight number or log
slot to int32_t.

+++ src/kernel/ao_tracker.c
@@ -186,9 +186,9 @@ ao_gps_lock(void)
 static uint8_t erasing_current;
 
 void
-ao_tracker_erase_start(uint16_t flight)
+ao_tracker_erase_start(int32_t flight)
 {
-	erasing_current = flight == ao_flight_number;
+	erasing_current = flight == (int32_t) ao_flight_number;
 	if (erasing_current) {
 		ao_mutex_get(&tracker_mutex);
 		ao_log_stop();
